Fix fox logo password following
The new lightweight svg logo was not following text quite right.

The new `lookAt` method was not using the same logic the module was using internally on mouse movement.

I simply used that logic and exposed it via the old (expected) API, and got it behaving the way I like.