Version v10.30.2 RC (#19109)
* Version v10.30.2

* Ensure the GasDetailsItem component can handle a tx with a maxPriorityFee of 0 (#19102)

* Ensure the GasDetailsItem component can handle a tx with a maxPriorityFee of 0

* Clean up code

* Update ui/components/app/gas-details-item/gas-details-item.js

Co-authored-by: Mark Stacey <markjstacey@gmail.com>

---------

Co-authored-by: Mark Stacey <markjstacey@gmail.com>

* Backport the `@metamask/message-manager` update to v5 (#19110)

The `@metamask/message-manager` update to v5 has been backported to the
v10.30.x release branch. This update includes three breaking changes,
but two (bump in minimum supported Node.js version, and change in type
from `Map` to `Record`) don't affect this project. The only breaking
change requiring changes was the addition of the `getCurrentChainId`
constructor parameter for the TypedMessageManger.

* Update changelog

---------

Co-authored-by: MetaMask Bot <metamaskbot@users.noreply.github.com>
Co-authored-by: Dan J Miller <danjm.com@gmail.com>
Co-authored-by: Mark Stacey <markjstacey@gmail.com>