devDeps: @lavamoat/allow-scripts@2.0.3->2.3.1 (#20018)
* devDeps: @lavamoat/allow-scripts@2.0.3->2.3.1

Note: As of right now, this causes depcheck script to fail under yarn due to a server error in NPM registry triggered by https://github.com/yarnpkg/berry/issues/4117. As can be seen in that issue, the results from this script have already been invalid since the upgrade from yarnv1 to yarnv3 and a change of package manager version or dependency-auditing script will have to be made.